Huge feature update of the whole library, BC-BREAK!
* Greatly improved module documentation
* Code style improvements with the use of pylint tool
* BC breaking refactoring to unify code style
* Multiple minor bugfixes

Note: I am well aware of the ugliness of such a huge commit. However all of the changes were done in a short time and as part of a single big feature upgrade and in my opinion in this case I can justify that, because all changes are related to one another.