Mask PID in PID file name and runlog file name
Format it as hex to mask it due to the fact that in case of PID files
the PID in file name is not the same as PID within the file. The reason
is the fact that the PID file name is determined before daemonization.

Alternative approach would be to change the daemonization library to be
able to generate the PID file name. This however would restrict the
library too much.